requirements.txt
# Core pipeline
yt-dlp>=2024.0.0
faster-whisper>=1.0.0
scenedetect[opencv]>=0.6.0
numpy>=1.26.0
librosa>=0.10.0

# Claude API
anthropic>=0.60.0

# Video editing
ffmpeg-python>=0.2.0  # NOTE: upstream abandoned since 2020 — works but no active maintenance
pysubs2>=1.7.0
Pillow>=12.0.0

# Config & CLI
pyyaml>=6.0
click>=8.0.0
python-dotenv>=1.0.0

# Web API (FastAPI stack)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
jinja2>=3.1.0
sse-starlette>=3.0.0
httpx>=0.27.0

# Async & background
aiosqlite>=0.20.0
aiofiles>=23.0.0
apscheduler>=3.10.0
alembic>=1.13.0

# Reliability
rich>=14.0.0
tenacity>=9.0.0
